source/requirements.txt
django>=4.2
djangorestframework
psycopg2-binary
docker
gunicorn==23.0.0
channels
redis
python-dotenv